+++ client/service/service.go
@@ -28,10 +28,9 @@ import (
 	"github.com/sagernet/sing/service"
 	"github.com/sagernet/sing/service/pause"
 
+	"github.com/getlantern/radiance/config"
 	"github.com/getlantern/sing-box-extensions/protocol"
 	"github.com/getlantern/sing-box-extensions/ruleset"
-
-	"github.com/getlantern/radiance/config"
 )
 
 // BoxService is a wrapper around libbox.BoxService
